feat: implement monster deaths
