+++ rustyhack_server/src/game/combat.rs
@@ -32,17 +32,17 @@ pub(crate) type Defender = Uuid;
 pub(crate) type CombatAttackerStats = HashMap<Uuid, Stats>;
 
 pub(crate) fn resolve_combat(attacker_stats: &Stats, defender_stats: &Stats) -> f32 {
-    info!("Resolving combat...");
+    debug!("Resolving combat...");
     if check_attack_success(attacker_stats.dex, defender_stats.dex) {
-        info!("Attack hit...");
+        debug!("Attack hit...");
         let actual_damage_received = calculate_actual_damage_received(
             calculate_damage_dealt(BASE_WEAPON_DAMAGE, attacker_stats.str),
             defender_stats.armour,
         );
-        info!("Damage taken: {}", actual_damage_received);
+        debug!("Damage taken: {}", actual_damage_received);
         actual_damage_received
     } else {
-        info!("Attack missed...");
+        debug!("Attack missed...");
         0.0
     }
 }
@@ -50,7 +50,7 @@ pub(crate) fn resolve_combat(attacker_stats: &Stats, defender_stats: &Stats) -> 
 fn calculate_damage_dealt(attacker_weapon_damage_range: Range<f32>, attacker_str: f32) -> f32 {
     let mut rng = thread_rng();
     let attacker_weapon_damage = rng.gen_range(attacker_weapon_damage_range);
-    info!(
+    debug!(
         "Weapon damage before strength modifier: {}",
         attacker_weapon_damage
     );
