+++ rustyhack_server/src/game/systems/player_systems.rs
@@ -1,3 +1,4 @@
+use crate::game::players::PlayersPositions;
 use legion::world::SubWorld;
 use legion::{system, Query};
 use rustyhack_lib::ecs::components::{PlayerDetails, Position, Stats};
@@ -14,5 +15,17 @@ pub(crate) fn resolve_player_deaths(
             *position = Position::default();
             position.update_available = true;
         }
+    }
+}
+
+#[system]
+pub(crate) fn update_player_positions_resource(
+    world: &mut SubWorld,
+    query: &mut Query<(&PlayerDetails, &Position)>,
+    #[resource] players_positions: &mut PlayersPositions,
+) {
+    players_positions.clear();
+    for (player_details, position) in query.iter(world) {
+        players_positions.insert(player_details.id, position.clone());
     }
 }
